# docs/specs/api-specs/openapi-missing-endpoints.yaml
openapi: 3.0.3
info:
  title: Cortex API - Missing Endpoints
  version: 1.0.0
  description: OpenAPI specification for missing and incomplete API endpoints
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.cortex.example.com/api
    description: Production server
...
paths:
  /projects/{projectId}/ingest/jobs/{jobId}:
    delete:
      tags:
        - ingest
      summary: Delete ingest job
      operationId: deleteIngestJob
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/JobId'
      responses:
        '204':
          description: Job successfully deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []

    get:
      tags:
        - ingest
      summary: Get ingest job
      operationId: getIngestJob
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestJob'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []
  /projects/{projectId}/ingest/jobs/{jobId}/cancel:
    post:
      tags:
        - ingest
      summary: Cancel ingest job
      operationId: cancelIngestJob
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []
...
  /projects/{projectId}/agent-runs/{runId}:
    get:
      tags:
        - agents
      summary: Get agent run
      operationId: getAgentRun
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRun'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []
  /projects/{projectId}/agent-runs/{runId}/steps:
    get:
      tags:
        - agents
      summary: List agent run steps
      operationId: listAgentRunSteps
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: runId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: List of steps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAgentSteps'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []
  /projects/{projectId}/agent-runs/{runId}/messages:
    get:
      tags:
        - agents
      summary: List agent run messages
      operationId: listAgentRunMessages
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: runId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAgentMessages'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []

    post:
      tags:
        - agents
      summary: Append message to agent run
      operationId: appendAgentRunMessage
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: runId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppendMessageRequest'
      responses:
        '201':
          description: Message appended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []
  /projects/{projectId}/agent-runs/{runId}/cancel:
    post:
      tags:
        - agents
      summary: Cancel agent run
      operationId: cancelAgentRun
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Run cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRun'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: []
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
      description: Project ID
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
      description: Ingest job ID
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Pagination cursor
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Maximum number of results
  schemas:
    IngestJob:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        status:
          type: string
          enum: [QUEUED, RUNNING, COMPLETED, FAILED, CANCELLED]
        progress:
          type: number
          minimum: 0
          maximum: 1
...
    AgentRun:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        workflowId:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED, CANCELLED]
    AgentStep:
      type: object
      properties:
        id:
          type: string
        runId:
          type: string
        status:
          type: string
    AgentMessage:
      type: object
      properties:
        id:
          type: string
        runId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
    PaginatedAgentSteps:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AgentStep'
        nextCursor:
          type: string
          nullable: true
        total:
          type: integer
    PaginatedAgentMessages:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AgentMessage'
        nextCursor:
          type: string
          nullable: true
        total:
          type: integer
    AppendMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
        contextItemIds:
          type: array
          items:
            type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: 'openapi-error-responses.yaml#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: 'openapi-error-responses.yaml#/components/schemas/Error'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []